+++ models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import product_bundle
+from . import product_bundle_line
